// featurehub-javascript-client-sdk/app/models/models/feature-environment-collection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FeatureServiceApi
 * This describes the API clients use for accessing features
 *
 * The version of the OpenAPI document: 1.1.8
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { FeatureState } from "./feature-state";

/**
 * This represents a collection of features as per a request from a GET api. GET\'s can request multiple API Keys at the same time.
 * @export
 * @interface FeatureEnvironmentCollection
 */
export interface FeatureEnvironmentCollection {
  /**
   *
   * @type {string}
   * @memberof FeatureEnvironmentCollection
   */
  id: string;
  /**
   *
   * @type {Array<FeatureState>}
   * @memberof FeatureEnvironmentCollection
   */
  features?: Array<FeatureState>;
}