// featurehub-javascript-client-sdk/app/models/models/feature-value-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FeatureServiceApi
 * This describes the API clients use for accessing features
 *
 * The version of the OpenAPI document: 1.1.8
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */
export enum FeatureValueType {
    Boolean = 'BOOLEAN',
    String = 'STRING',
    Number = 'NUMBER',
    Json = 'JSON'
}